Rename files and separate implementations to different files
